[example-studio] Make example with custom blocks
